pull up-to-date currency data on the fly
The system is intended to give accurate results based on a selected base
currency (currently DKK).
When updateCurrency is first called it will install default values into
the base->* table and divide out derived values into all the other
currency conversion tables. It will then kick off an XHR to fetch
updated values.
When the updated values have been fetched it will perform the same
procedure as on initialization.
If the user selects another currency in the dropdown, the XHR will be
kicked off to fetch the new currency (if not already fetched) and that
currency's table will be updated with accurate values (rather than the
divided out ones we calculated earlier).

All currency functions with the exception of the init/update stuff now
takes an explicit parameter that specifies the desired currency.

The issue with this is of course that you might get very slightly
different results when switching currencies. I doubt anyone will [be
able to] notice.