currency: fix init race
I'm pretty sure the threading model of JS is "no.", so we don't need to
concern ourselves with atomicity. Here though we have the problem that
two currency fetches might be kicked off in quick succession and the
responses coming back out of order, so we need to make doubly sure that
only the first response kicks off the recalculation of the approximated
currency map.